+++ bin/generate.dart
@@ -1,5 +1,0 @@
-import 'package:app_sizer/src/helper/generate_prescale.dart' as generator;
-
-void main(List<String> args) {
-  generator.main(args);
-}